+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/ProblemSpec.java
@@ -109,9 +109,9 @@ public interface ProblemSpec {
      * @param path the file location
      * @param line the one-indexed line number
      * @return this
-     * @since 8.7
+     * @since 8.6
      */
-    ProblemSpec lineInFileLocation(String path, long line);
+    ProblemSpec lineInFileLocation(String path, int line);
 
     /**
      * Declares that this problem is in a file with on a line at a certain position.
@@ -121,9 +121,9 @@ public interface ProblemSpec {
      * @param line the one-indexed line number
      * @param column the one-indexed column
      * @return this
-     * @since 8.7
+     * @since 8.6
      */
-    ProblemSpec lineInFileLocation(String path, long line, long column);
+    ProblemSpec lineInFileLocation(String path, int line, int column);
 
     /**
      * Declares that this problem is in a file with on a line at a certain position.
@@ -133,9 +133,9 @@ public interface ProblemSpec {
      * @param column the one-indexed column
      * @param length the length of the text
      * @return this
-     * @since 8.7
+     * @since 8.6
      */
-    ProblemSpec lineInFileLocation(String path, long line, long column, long length);
+    ProblemSpec lineInFileLocation(String path, int line, int column, int length);
 
     /**
      * Declares that this problem is in a file at a certain global position with a given length.
@@ -144,9 +144,9 @@ public interface ProblemSpec {
      * @param offset the zero-indexed global offset from the beginning of the file
      * @param length the length of the text
      * @return this
-     * @since 8.7
+     * @since 8.6
      */
-    ProblemSpec offsetInFileLocation(String path, long offset, long length);
+    ProblemSpec offsetInFileLocation(String path, int offset, int length);
 
     /**
      * Declares that this problem is emitted while applying a plugin.
